refactor(RestaurantTab): restaurant-tab-body 컴포넌트 내 body 부분을 이동하는 형태로 변경
